feat: extract @pop-choice/shared package with common service utilities
- Add packages/shared with logger (pino), db pool + ops, embeddings, and config helpers
- Services now import from @pop-choice/shared instead of duplicating implementations
- movie-backfill retains its own DB operations (getIncompleteMovies, updateMovie)
- checkTableExists moved to shared as a generic utility
- Remove openai, pg, pino direct deps from services (provided by shared)
- Update .gitignore: remove redundant services/*/node_modules and services/*/dist entries
