+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "type": "module",
   "workspaces": [
+    "packages/*",
     "services/*"
   ],
   "scripts": {
