+++ services/movie-backfill/package.json
@@ -10,8 +10,7 @@
     "test": "vitest run"
   },
   "dependencies": {
-    "openai": "^6.33.0",
-    "pg": "^8.20.0",
+    "@pop-choice/shared": "*",
     "zod": "^4.3.6"
   },
   "devDependencies": {
