+++ services/movie-discovery/package.json
@@ -10,9 +10,8 @@
     "test": "vitest run"
   },
   "dependencies": {
-    "node-cron": "^4.2.1",
-    "openai": "^6.34.0",
-    "pg": "^8.20.0"
+    "@pop-choice/shared": "*",
+    "node-cron": "^4.2.1"
   },
   "devDependencies": {
     "@types/node": "^24.12.2",
