+++ services/movie-discovery/src/config.ts
@@ -1,3 +1,5 @@
+import { parsePositiveFloat, parsePositiveInt, requireEnvVars } from '@pop-choice/shared';
+
 export type TMDBSource = 'now_playing' | 'upcoming' | 'top_rated' | 'popular';
 
 export interface Config {
@@ -14,39 +16,8 @@ export interface Config {
   schedule: string;
 }
 
-function parsePositiveInt(value: string | undefined, defaultValue: number, name: string): number {
-  if (value === undefined || value === '') return defaultValue;
-  const num = Number(value);
-  if (!Number.isInteger(num) || num <= 0) {
-    throw new Error(`${name} must be a positive integer, got: ${JSON.stringify(value)}`);
-  }
-  return num;
-}
-
-function parsePositiveFloat(value: string | undefined, defaultValue: number, name: string): number {
-  if (value === undefined || value === '') return defaultValue;
-  const num = Number(value);
-  if (!Number.isFinite(num) || num <= 0) {
-    throw new Error(`${name} must be a positive number, got: ${JSON.stringify(value)}`);
-  }
-  return num;
-}
-
 export function loadConfig(): Config {
-  const missing: string[] = [];
-
-  const tmdbApiKey = process.env.TMDB_API_KEY;
-  if (!tmdbApiKey) missing.push('TMDB_API_KEY');
-
-  const openaiApiKey = process.env.OPENAI_API_KEY;
-  if (!openaiApiKey) missing.push('OPENAI_API_KEY');
-
-  const databaseUrl = process.env.DATABASE_URL;
-  if (!databaseUrl) missing.push('DATABASE_URL');
-
-  if (missing.length > 0) {
-    throw new Error(`Missing required environment variables: ${missing.join(', ')}`);
-  }
+  const env = requireEnvVars(['TMDB_API_KEY', 'OPENAI_API_KEY', 'DATABASE_URL']);
 
   const validSources: TMDBSource[] = ['now_playing', 'upcoming', 'top_rated', 'popular'];
   const rawSources = process.env.TMDB_SOURCES?.split(',').map((s) => s.trim()) ?? [];
@@ -60,13 +31,11 @@ export function loadConfig(): Config {
     );
   }
 
-  const sources: TMDBSource[] = rawSources.length > 0 ? filteredSources : validSources;
-
   return {
-    tmdbApiKey: tmdbApiKey!,
-    openaiApiKey: openaiApiKey!,
-    databaseUrl: databaseUrl!,
-    sources,
+    tmdbApiKey: env.TMDB_API_KEY,
+    openaiApiKey: env.OPENAI_API_KEY,
+    databaseUrl: env.DATABASE_URL,
+    sources: rawSources.length > 0 ? filteredSources : validSources,
     maxPagesPerSource: parsePositiveInt(
       process.env.MAX_PAGES_PER_SOURCE,
       3,
@@ -77,6 +46,6 @@ export function loadConfig(): Config {
     maxMoviesPerRun: parsePositiveInt(process.env.MAX_MOVIES_PER_RUN, 50, 'MAX_MOVIES_PER_RUN'),
     language: process.env.TMDB_LANGUAGE?.trim() || 'en-US',
     dryRun: process.env.DRY_RUN === 'true',
-    schedule: process.env.SYNC_SCHEDULE?.trim() ?? '0 0 * * 0', // Default: weekly Sunday midnight UTC
+    schedule: process.env.SYNC_SCHEDULE?.trim() ?? '0 0 * * 0',
   };
 }
