+++ services/movie-seed/package.json
@@ -9,9 +9,7 @@
     "start": "node dist/index.js"
   },
   "dependencies": {
-    "openai": "^6.33.0",
-    "pg": "^8.20.0",
-    "pino": "^10.3.1"
+    "@pop-choice/shared": "*"
   },
   "devDependencies": {
     "@types/node": "^24.12.2",
